Module bin_data::context

Context for encoding and decoding.

The types can be roughly grouped into two categories:

Context::EndianContext: explicit endianness specification and inheritance.

Endian: little-endian or big-endian.

NoEndian: endianness is not decided at runtime.

Context::ArgsBuilder: type-level construct for named arguments.

NoArgs: no argument at all, or Args = ().

VecArgs and VecArgsBuilder: arguments for Vec, slices, etc.

StrArgs and StrArgsBuilder: arguments for String, str, etc.

Types in this module might appear in error messages, here is an overview:

expected enum Endian, found struct NoEndian: endianness for one of the fields must be specified. Add the following attribute: #[bin_data(endian = "...")], either to the whole struct, or to individual fields.

expected struct NoEndian, found enum Endian: opposite situation to the previous one, one of the fields require no runtime-specified endianness, but endianness is specified explicitly using #[bin_data(endian = "...")]. Remove that superfluous attribute.

some argument builder does not implement ArgsBuilderFinished: some Required argument is not specified. Specify it in #[bin_data(args { ... })]. The arguments should be given as name = value, separated and optionally ended by commas. Under the hood, it calls the method name with argument value on the argument builder. For instance, the following code set the expected length of a Vec by calling VecArgsBuilder::count:

#[bin_data(args:decode { count = 42 })]
xs: Vec<u8>,

See also each type’s documentation for detailed explanation.

Structs

NoArgs

Trivial arguments builder for ().

NoEndian

Indicate the endianness is not determined at runtime.

Provided

A provided field. This is deliberately not Default, to prevent accidentally supplying default values for required arguments.

Required

Placeholder for a required field without a default value.

StrArgs

Arguments for encoding or decoding a str, String, etc.

StrArgsBuilder

Named arguments builder for StrArgs.

VecArgs

Arguments for encoding or decoding a Vec.

VecArgsBuilder

Named arguments builder for VecArgs.

Enums

Endian

Endianness for integers, floating-point numbers, etc.

Traits

ArgsBuilderFinished

Indicates that all arguments is supplied.

Context

Specify the named argument builder and endianness context.