Trait bin_data::context::Context

pub trait Context<Dir: Direction> {
    type EndianContext: EndianContext;
    type ArgsBuilder;

    fn args_builder() -> Self::ArgsBuilder;

    fn args_builder_of_val(&self) -> Self::ArgsBuilder { ... }
}

Specify the named argument builder and endianness context.

Required Associated Types

type EndianContext: EndianContext

Context containing the desired endianness.

type ArgsBuilder

The argument builder type.

Required Methods

fn args_builder() -> Self::ArgsBuilder

Create an argument builder with default settings.

Provided Methods

fn args_builder_of_val(&self) -> Self::ArgsBuilder

Create an argument builder with default settings. Types of the values specified in #[bin_data(encode = ...)] is unknown to the bin_data macro, so we use this method to let the compiler infer a proper Self type and Self::ArgsBuilder.