Trait bin_data::data::Encode

pub trait Encode<Args = ()>: Context<Write> {
    fn encode_with<W: Write + ?Sized>(
        &self,
        writer: &mut W,
        endian: Self::EndianContext,
        args: Args
    ) -> Result<(), EncodeError>;

    fn encode<W: Write + ?Sized>(
        &self,
        writer: &mut W
    ) -> Result<(), EncodeError>
    where
        Self::EndianContext: Default,
        Self::ArgsBuilder: ArgsBuilderFinished<Output = Args>,
    { ... }
}

Encode binary data from structured in-memory representation.

Required Methods

fn encode_with<W: Write + ?Sized>(
&self,
writer: &mut W,
endian: Self::EndianContext,
args: Args
) -> Result<(), EncodeError>

Encode self to the output stream with the given arguments.

Provided Methods

fn encode<W: Write + ?Sized>(&self, writer: &mut W) -> Result<(), EncodeError>where
Self::EndianContext: Default,
Self::ArgsBuilder: ArgsBuilderFinished<Output = Args>,

Encode self to the output stream with default arguments.